🚚 move DD package code to source file
